requirements.txt
telebot
requests
SpeechRecognition
gtts
python-decouple